chore: stripe sync cleanup
