+++ server/src/internal/billing/v2/providers/stripe/utils/sync/fetchStripeSyncObjects.ts
@@ -1,0 +1,31 @@
+import type Stripe from "stripe";
+import {
+	STRIPE_SYNC_SCHEDULE_EXPAND,
+	STRIPE_SYNC_SUBSCRIPTION_EXPAND,
+} from "./stripeItemSnapshot/stripeSyncExpand";
+
+export const fetchStripeSyncSubscription = async ({
+	stripeCli,
+	subscriptionId,
+}: {
+	stripeCli: Stripe;
+	subscriptionId?: string;
+}) =>
+	subscriptionId
+		? stripeCli.subscriptions.retrieve(subscriptionId, {
+				expand: STRIPE_SYNC_SUBSCRIPTION_EXPAND,
+			})
+		: null;
+
+export const fetchStripeSyncSchedule = async ({
+	stripeCli,
+	scheduleId,
+}: {
+	stripeCli: Stripe;
+	scheduleId?: string;
+}) =>
+	scheduleId
+		? stripeCli.subscriptionSchedules.retrieve(scheduleId, {
+				expand: STRIPE_SYNC_SCHEDULE_EXPAND,
+			})
+		: null;
